refactor: organize engine domains
