+++ src/engine/core/gameEngine.ts
@@ -1,23 +1,23 @@
 import { fatalError, logDebug } from '@utils/logMessage'
 import { MessageBus, type IMessageBus } from '@utils/messageBus'
 import type { ILoader } from '@loader/loader'
-import { END_TURN_MESSAGE, ENGINE_STATE_CHANGED_MESSAGE, MAP_SWITCHED_MESSAGE, POSITION_CHANGED_MESSAGE, SWITCH_PAGE_MESSAGE } from './messages'
+import { END_TURN_MESSAGE, ENGINE_STATE_CHANGED_MESSAGE, MAP_SWITCHED_MESSAGE, POSITION_CHANGED_MESSAGE, SWITCH_PAGE_MESSAGE } from '../dialog/messages'
 import type { IStateManager } from './stateManager'
 import { TrackedValue, type ITrackedValue } from '@utils/trackedState'
-import type { ITranslationService } from './translationService'
-import type { IPageManager } from './pageManager'
+import type { ITranslationService } from '../dialog/translationService'
+import type { IPageManager } from '../page/pageManager'
 import type { Action } from '@loader/data/action'
 import type { CleanUp } from '@utils/types'
-import type { IMapManager } from './mapManager'
-import type { IVirtualInputHandler } from './virtualInputHandler'
-import type { IInputManager } from './inputManager'
-import type { IScriptRunner, ScriptContext } from './scriptRunner'
+import type { IMapManager } from '../map/mapManager'
+import type { IVirtualInputHandler } from '../input/virtualInputHandler'
+import type { IInputManager } from '../input/inputManager'
+import type { IScriptRunner, ScriptContext } from '../script/scriptRunner'
 import type { Condition } from '@loader/data/condition'
-import type { IOutputManager } from './outputManager'
-import type { IDialogManager } from './dialogManager'
+import type { IOutputManager } from '../output/outputManager'
+import type { IDialogManager } from '../dialog/dialogManager'
 import type { ContextData } from './context'
-import type { IActionHandler } from './actions/actionHandler'
-import type { IConditionResolver } from './conditions/conditionResolver'
+import type { IActionHandler } from '../actions/actionHandler'
+import type { IConditionResolver } from '../conditions/conditionResolver'
 
 export const GameEngineState = {
     init: 0,
