+++ src/engine/core/gameEngineInitializer.ts
@@ -3,16 +3,16 @@ import { MessageBus } from '@utils/messageBus'
 import { ChangeTracker } from './changeTracker'
 import { StateManager, type IStateManager } from './stateManager'
 import type { ContextData } from './context'
-import type { IPageManager } from './pageManager'
-import type { IMapManager } from './mapManager'
-import type { IVirtualInputHandler } from './virtualInputHandler'
-import type { IInputManager } from './inputManager'
-import type { IOutputManager } from './outputManager'
-import type { IDialogManager } from './dialogManager'
-import type { ITranslationService } from './translationService'
-import type { IScriptRunner } from './scriptRunner'
-import type { IActionHandler } from './actions/actionHandler'
-import type { IConditionResolver } from './conditions/conditionResolver'
+import type { IPageManager } from '../page/pageManager'
+import type { IMapManager } from '../map/mapManager'
+import type { IVirtualInputHandler } from '../input/virtualInputHandler'
+import type { IInputManager } from '../input/inputManager'
+import type { IOutputManager } from '../output/outputManager'
+import type { IDialogManager } from '../dialog/dialogManager'
+import type { ITranslationService } from '../dialog/translationService'
+import type { IScriptRunner } from '../script/scriptRunner'
+import type { IActionHandler } from '../actions/actionHandler'
+import type { IConditionResolver } from '../conditions/conditionResolver'
 import { TurnScheduler } from './turnScheduler'
 import { GameEngine, type IGameEngine } from './gameEngine'
