+++ src/engine/core/stateManager.ts
