+++ src/engine/dialog/dialogManagerService.ts
@@ -1,4 +1,4 @@
-import type { IGameEngine } from './gameEngine'
+import type { IGameEngine } from '../core/gameEngine'
 import { DialogManager, type IDialogManager } from './dialogManager'
 import type { IMessageBus } from '@utils/messageBus'
