+++ src/engine/dialog/messages.ts
@@ -11,3 +11,4 @@ export const ADD_LINE_TO_OUTPUT_LOG = 'SYSTEM.ADD-LINE-TO-OUTPUT-LOG'
 export const OUTPUT_LOG_LINE_ADDED = 'SYSTEM.OUTPUT-LOG-LINE-ADDED'
 export const START_DIALOG = 'SYSTEM.START-DIALOG'
 export const DIALOG_STARTED = 'SYSTEM.DIALOG-STARTED'
+
