+++ src/engine/page/pageManager.ts
@@ -1,9 +1,9 @@
 import { logDebug } from '@utils/logMessage'
 import type { ILoader } from '@loader/loader'
 import type { IMessageBus } from '@utils/messageBus'
-import type { IStateManager } from './stateManager'
-import type { ContextData } from './context'
-import { PAGE_SWITCHED_MESSAGE, SWITCH_PAGE_MESSAGE } from './messages'
+import type { IStateManager } from '../core/stateManager'
+import type { ContextData } from '../core/context'
+import { PAGE_SWITCHED_MESSAGE, SWITCH_PAGE_MESSAGE } from '../dialog/messages'
 
 export interface IPageManager {
     initialize(): void
