+++ src/main.tsx
@@ -1,19 +1,19 @@
 import { createRoot } from 'react-dom/client'
 import { logDebug } from '@utils/logMessage'
 import { Loader, type ILoader } from '@loader/loader'
-import { type IGameEngine } from '@engine/gameEngine'
-import { GameEngineInitializer, type IEngineManagerFactory } from '@engine/gameEngineInitializer'
+import { type IGameEngine } from '@engine/core/gameEngine'
+import { GameEngineInitializer, type IEngineManagerFactory } from '@engine/core/gameEngineInitializer'
 import { PostMessageActionHandler } from '@engine/actions/postMessageActionHandler'
 import { ScriptActionHandler } from '@engine/actions/scriptActionHandler'
 import { ScriptConditionResolver } from '@engine/conditions/scriptConditionResolver'
-import { createPageManager } from '@engine/pageManagerService'
-import { createMapManager } from '@engine/mapManagerService'
-import { createVirtualInputHandler } from '@engine/virtualInputHandlerService'
-import { createInputManager } from '@engine/inputManagerService'
-import { createOutputManager } from '@engine/outputManagerService'
-import { createDialogManager } from '@engine/dialogManagerService'
-import { createScriptRunner } from '@engine/scriptRunnerFactory'
-import { createTranslationService } from '@engine/translationServiceFactory'
+import { createPageManager } from '@engine/page/pageManagerService'
+import { createMapManager } from '@engine/map/mapManagerService'
+import { createVirtualInputHandler } from '@engine/input/virtualInputHandlerService'
+import { createInputManager } from '@engine/input/inputManagerService'
+import { createOutputManager } from '@engine/output/outputManagerService'
+import { createDialogManager } from '@engine/dialog/dialogManagerService'
+import { createScriptRunner } from '@engine/script/scriptRunnerFactory'
+import { createTranslationService } from '@engine/dialog/translationServiceFactory'
 import { App } from '@app/app'
 import './style/reset.css'
 import './style/variables.css'
