+++ test/engine/stateManager.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
-import { ChangeTracker } from '@engine/changeTracker'
-import { StateManager } from '@engine/stateManager'
+import { ChangeTracker } from '@engine/core/changeTracker'
+import { StateManager } from '@engine/core/stateManager'
 
 interface Data extends Record<string, unknown> { count: number }
